[1주차-step2] refactor(station): package 분리
